+++ src/main/kotlin/dsm/pick2024/infrastructure/feign/neis/dto/response/NeisFeignClientTimetableResponse.kt
@@ -1,4 +1,4 @@
-package dsm.pick2024.infrastructure.feign.neis.dto.request
+package dsm.pick2024.infrastructure.feign.neis.dto.response
 
 data class NeisFeignClientTimetableResponse(
     val hisTimetable: List<HisTimetable>
